Implement LZ4 compression

+++ write_pkg.py
@@ -1,4 +1,5 @@
-# ED8 .pkg builder.  Currently does not apply any compression.
+# ED8 .pkg builder.  The lz4 module must be present for compression (which is
+# specified at the command line).
 #
 # GitHub eArmada8/ed8pkg2gltf
 
@@ -34,7 +35,13 @@ def write_pkg_file (newfilename, file_stream, content_struct, magic = b'\x00\x00
         f.write(file_stream.read())
     return
 
-def processFolder(pkg_folder):
+def processFolder(pkg_folder, lz4_compress = False, overwrite = False):
+    if lz4_compress == True:
+        try:
+            import lz4.block
+            lz4_present = True
+        except ModuleNotFoundError:
+            lz4_present = False
     if os.path.exists(pkg_folder):
         files = glob.glob(pkg_folder+'/*.*')
         f = io.BytesIO()
@@ -42,13 +49,25 @@ def processFolder(pkg_folder):
         for i in range(len(files)):
             with open(files[i], 'rb') as f2:
                 binary_file_data = f2.read()
-                file_details = {"file_entry_name": os.path.basename(files[i]), "file_entry_uncompressed_size": len(binary_file_data),\
+                unc_size = len(binary_file_data)
+                flags = 0
+                if lz4_compress == True:
+                    if lz4_present == True:
+                        binary_file_data = lz4.block.compress(binary_file_data, mode = 'high_compression', store_size=False)
+                        flags = 4
+                file_details = {"file_entry_name": os.path.basename(files[i]), "file_entry_uncompressed_size": unc_size,\
                     "file_entry_compressed_size": len(binary_file_data), "file_entry_offset": 0,\
-                    "file_entry_flags": 0}
+                    "file_entry_flags": flags}
                 content_struct = insert_file_into_stream (f, content_struct, binary_file_data, file_details)
-        if os.path.exists("{0}.pkg".format(pkg_folder)):
-            shutil.copy2("{0}.pkg".format(pkg_folder), "{0}.pkg.bak".format(pkg_folder))
-        write_pkg_file ("{0}.pkg".format(pkg_folder), f, content_struct, magic = b'\x00\x00\x00\x00')
+        if os.path.exists("{0}.pkg".format(pkg_folder)) and (overwrite == False):
+            if str(input("{0}.pkg".format(pkg_folder) + " exists! Overwrite? (y/N) ")).lower()[0:1] == 'y':
+                overwrite = True
+        if (overwrite == True) or not os.path.exists("{0}.pkg".format(pkg_folder)):
+            if os.path.exists("{0}.pkg".format(pkg_folder)):
+                shutil.copy2("{0}.pkg".format(pkg_folder), "{0}.pkg.bak".format(pkg_folder))
+            if lz4_compress == True and lz4_present == False:
+                print("LZ4 compression specified but module is not installed, .pkg not compressed.")
+            write_pkg_file ("{0}.pkg".format(pkg_folder), f, content_struct, magic = b'\x00\x00\x00\x00')
 
 if __name__ == "__main__":
     # Set current directory
@@ -58,6 +77,8 @@ if __name__ == "__main__":
         import argparse
         parser = argparse.ArgumentParser()
         parser.add_argument('pkg_folder', help="Name of folder to pack.")
+        parser.add_argument('-l', '--lz4_compress', help="Apply LZ4 compression", action="store_true")
+        parser.add_argument('-o', '--overwrite', help="Overwrite existing files", action="store_true")
         args = parser.parse_args()
         if os.path.exists(args.pkg_folder):
-            processFolder(args.pkg_folder)
+            processFolder(args.pkg_folder, lz4_compress = args.lz4_compress, overwrite = args.overwrite)
